pyproject.toml
[project]
name = "smart-docs"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "fastapi[all] (>=0.120.1,<0.121.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "openai (>=2.6.0,<3.0.0)",
    "accelerate (>=1.11.0,<2.0.0)",
    "huggingface-hub (>=0.36.0,<0.37.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "chromadb (>=1.2.2,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.0.1,<8.0.0)"
]
